Fix bug in ordering from processing older messages.
The old message processing had a small bug where each chunk of
messages ended up in reversed order.  For example if you had an
ordering like:

ABCDEF

and change the filtering, they might be reordered as

CBAFED

This was a small bug that is fixed by correctly building the early
messages queue in the proper reversed order.