Fixing issue with master disconnect/reconnect.
After merging in ros_thread, I noticed that the console no longer
reconnected properly when if the ros core is shutdown and restarted.
I think this was due to using ros::Rate to limit the main RosThread
loop.  I removed this in favor of QThread::msleep so that the loop is
independent of ROS.  Also removed the added ros::NodeHandle member and
initial startRos call that I'm guessing was added to allow ros::Rate
to run without an exception.

Using QThread::msleep might be be preferrable anyways so that the
console continues to read messages immediately even if we're running
simulations with the ROS clock turned waaaay down.