Fix issue 05 - Node list not clearing out properly.
The root cause of this problem was that the message count in the log
database was not being cleared out.  I've refactored the structure a
little bit to make the dependencies a little cleaner.  The log
database doesn't know about the NodeListModel anymore (though the
model now knows about the database).  This puts things more in line
with the log database just being the main source of information, and
the node list model and log database proxy being different views onto
the database.