Sync behavior between saving to bags and text files.
This commit closes #10 by using the same behavior when saving to text
and bag files.  Both now save the same messages that are in the user's
current view.  Partly because this lets the user save subsets or
everything by deliberately selecting everything, and partly because it
was a lot easier to change the save to bag behavior than vice versa.