Add ability to read either `/rosout` or `/rosout_agg` from bag files
Prefer `/rosout` if it exists, if not, fall back to `/rosout_agg`. Also, if no log messages are found in a bag emit a
warning to indicate to user the reason why log messages did not change.

Tested this with bag files containing only `/rosout`, only `/rosout_agg`, both `/rosout` and `/rosout_agg`, and neither
`/rosout` nor `/rosout_agg`.